refactor to make Swap file clean
